Updating our fork master from the original

+++ suffix.h
@@ -31,9 +31,11 @@
 
 #ifndef __ARGCONFIG_SUFFIX_H__
 
+#include <inttypes.h>
+
 const char *suffix_si_get(double *value);
 const char *suffix_binary_get(long long *value);
 const char *suffix_dbinary_get(double *value);
-long long suffix_binary_parse(const char *value);
+uint64_t suffix_binary_parse(const char *value);
 
 #endif
